+++ src/main/java/com/roommanagement/dto/response/JwtResponse.java
@@ -6,12 +6,12 @@ public class JwtResponse {
 	private String token;
 	private String type = "Bearer";
 	private String refreshToken;
-	private Long id;
+	private Integer id;
 	private String username;
 	private String email;
 	private List<String> roles;
 
-	public JwtResponse(String accessToken, String refreshToken, Long id, String username, String email, List<String> roles) {
+	public JwtResponse(String accessToken, String refreshToken, Integer id, String username, String email, List<String> roles) {
 		this.token = accessToken;
 		this.refreshToken = refreshToken;
 		this.id = id;
@@ -36,11 +36,11 @@ public class JwtResponse {
 		this.type = tokenType;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
