+++ src/main/java/com/roommanagement/repository/expense/ExpenseRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
-  List<Expense> findAllByDefaultFlagIsTrueAndUserId(Long userId);
-  List<Expense> findAllByDefaultFlagIsFalseAndUserId(Long userId);
+  List<Expense> findAllByDefaultFlagIsTrueAndUserId(Integer userId);
+  List<Expense> findAllByDefaultFlagIsFalseAndUserId(Integer userId);
 }
