+++ src/main/java/com/roommanagement/service/room/CommandRoomService.java
@@ -5,9 +5,9 @@ import com.roommanagement.dto.request.room.UpdateExpenseRequest;
 
 public interface CommandRoomService {
   void createRoom(CreateRoomRequestDto roomRequestDto, int quantity);
-  void updateRoom(CreateRoomRequestDto roomRequestDto, long roomId);
+  void updateRoom(CreateRoomRequestDto roomRequestDto, Integer roomId);
 
-  void deleteRoom(long roomId);
-  void rollbackRoom(long roomId);
-  void updateExpense(Long roomId, Integer expenseId, UpdateExpenseRequest updateExpenseRequest);
+  void deleteRoom(Integer roomId);
+  void rollbackRoom(Integer roomId);
+  void updateExpense(Integer roomId, Integer expenseId, UpdateExpenseRequest updateExpenseRequest);
 }
